Section2/pyproject.toml
[project]
name = "data-quality-monitoring"
version = "1.0.0"
description = "Data Quality Monitoring System for Grant Applications"
requires-python = ">=3.9"
dependencies = [
    "pyspark>=3.3.0",
    "python-json-logger>=4.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pandas>=1.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
# uv configuration

